changes with alfabet
