+++ src/database/alphabet.json
@@ -1,0 +1,189 @@
+{
+	"!.png": {
+		"frame": {"x":0,"y":270,"w":90,"h":90}
+	},
+
+	"#.png": {
+		"frame": {"x":130,"y":270,"w":90,"h":90}
+	},
+
+	"$.png": {
+		"frame": {"x":67,"y":0,"w":40,"h":40}
+	},
+
+	"%.png": {
+		"frame": {"x":90,"y":270,"w":40,"h":90}
+	},
+
+	"*.png": {
+		"frame": {"x":893,"y":180,"w":90,"h":90}
+	},
+
+	",.png": {
+		"frame": {"x":803,"y":180,"w":90,"h":90}
+	},
+
+	"-.png": {
+		"frame": {"x":623,"y":180,"w":90,"h":90}
+	},
+
+	"0.png": {
+		"frame": {"x":533,"y":180,"w":90,"h":90}
+	},
+
+	"1.png": {
+		"frame": {"x":220,"y":270,"w":90,"h":90}
+	},
+
+	"2.png": {
+		"frame": {"x":227,"y":0,"w":80,"h":80}
+	},
+
+	"3.png": {
+		"frame": {"x":467,"y":0,"w":80,"h":80}
+	},
+
+	"4.png": {
+		"frame": {"x":387,"y":0,"w":80,"h":80}
+	},
+
+	"5.png": {
+		"frame": {"x":147,"y":0,"w":80,"h":80}
+	},
+
+	"6.png": {
+		"frame": {"x":307,"y":0,"w":80,"h":80}
+	},
+
+	"7.png": {
+		"frame": {"x":707,"y":0,"w":80,"h":80}
+	},
+
+	"8.png": {
+		"frame": {"x":627,"y":0,"w":80,"h":80}
+	},
+
+	"9.png": {
+		"frame": {"x":547,"y":0,"w":80,"h":80}
+	},
+
+	"10.png": {
+		"frame": {"x":0,"y":0,"w":27,"h":24}
+	},
+
+	";.png": {
+		"frame": {"x":107,"y":0,"w":40,"h":60}
+	},
+
+	"?.png": {
+		"frame": {"x":490,"y":270,"w":40,"h":90}
+	},
+
+	"@.png": {
+		"frame": {"x":27,"y":0,"w":40,"h":40}
+	},
+
+	"A.png": {
+		"frame": {"x":710,"y":270,"w":90,"h":90}
+	},
+
+	"B.png": {
+		"frame": {"x":890,"y":270,"w":90,"h":90}
+	},
+
+	"C.png": {
+		"frame": {"x":450,"y":180,"w":83,"h":90}
+	},
+
+	"D.png": {
+		"frame": {"x":620,"y":270,"w":90,"h":90}
+	},
+
+	"E.png": {
+		"frame": {"x":530,"y":270,"w":90,"h":90}
+	},
+
+	"F.png": {
+		"frame": {"x":400,"y":270,"w":90,"h":90}
+	},
+
+	"G.png": {
+		"frame": {"x":310,"y":270,"w":90,"h":90}
+	},
+
+	"H.png": {
+		"frame": {"x":713,"y":180,"w":90,"h":90}
+	},
+
+	"I.png": {
+		"frame": {"x":180,"y":180,"w":90,"h":90}
+	},
+
+	"J.png": {
+		"frame": {"x":787,"y":0,"w":90,"h":90}
+	},
+
+	"K.png": {
+		"frame": {"x":90,"y":90,"w":90,"h":90}
+	},
+
+	"L.png": {
+		"frame": {"x":877,"y":0,"w":90,"h":90}
+	},
+
+	"M.png": {
+		"frame": {"x":0,"y":90,"w":90,"h":90}
+	},
+
+	"N.png": {
+		"frame": {"x":180,"y":90,"w":90,"h":90}
+	},
+
+	"O.png": {
+		"frame": {"x":360,"y":90,"w":90,"h":90}
+	},
+
+	"P.png": {
+		"frame": {"x":270,"y":90,"w":90,"h":90}
+	},
+
+	"Q.png": {
+		"frame": {"x":540,"y":90,"w":90,"h":90}
+	},
+
+	"R.png": {
+		"frame": {"x":360,"y":180,"w":90,"h":90}
+	},
+
+	"S.png": {
+		"frame": {"x":270,"y":180,"w":90,"h":90}
+	},
+
+	"T.png": {
+		"frame": {"x":90,"y":180,"w":90,"h":90}
+	},
+
+	"U.png": {
+		"frame": {"x":0,"y":180,"w":90,"h":90}
+	},
+
+	"V.png": {
+		"frame": {"x":630,"y":90,"w":90,"h":90}
+	},
+
+	"W.png": {
+		"frame": {"x":810,"y":90,"w":90,"h":90}
+	},
+
+	"X.png": {
+		"frame": {"x":720,"y":90,"w":90,"h":90}
+	},
+
+	"Y.png": {
+		"frame": {"x":450,"y":90,"w":90,"h":90}
+	},
+
+	"Z.png": {
+		"frame": {"x":800,"y":270,"w":90,"h":90}
+	}
+}
