+++ src/game-field.js
@@ -1,4 +1,5 @@
-import Spritesheet from './spritesheet';
+import Sprite from './Sprite';
+import Font from './Font';
 import {loadImage,loadJSON} from './loaders';
 import {createEntities} from './entities';
 import Timer from './timer';
@@ -12,14 +13,20 @@ export default function drawField(context, canvas) {
     Promise.all([
         loadImage('./images/sprites.png'),
         loadJSON('./sprites'),
-        loadJSON('./levels/1-2')
+        loadJSON('./levels/1-2'),
+        loadJSON('./alphabet'),
+        loadImage('./images/alphabet.png')
     ])
-    .then(([image, data, layout]) => {
+    .then(([tileImage, tileData, layout, fontData, fontsData]) => {
 
-        const sprites = new Spritesheet(image, data, canvas);
-
-        for (let sprite in data) {
+        const sprites = new Sprite(tileImage, tileData);
+        for (let sprite in tileData) {
             sprites.define(sprite);
+        }
+
+        const font = new Font(fontsData, fontData);
+        for (let letter in fontData) {
+            font.define(letter);
         }
 
         const drawBackgroundLayer = drawBackground(sprites, layout);
@@ -43,6 +50,11 @@ export default function drawField(context, canvas) {
 
                 drawBackgroundLayer(context);
                 sprites.update(deltaTime, context);
+                font.print('!@#$%*,', context, 0, 100, 90, 90);
+                font.print('1234567890', context, 0, 200, 90, 90);
+                font.print('QWERTYUIOP', context, 0, 300, 90, 90);
+                font.print('ASDFGHJKL', context, 0, 400, 90, 90);
+                font.print('ZXCVBNM', context, 0, 500, 90, 90);
 
                 drawCollisions(context, sprites.camera);
                 drawCameraView(context, sprites.camera);
